optimized global project layout and fixed some styles from /timeline page
