+++ app/layout.tsx
@@ -1,42 +1,20 @@
-'use client'
-
-import { Toaster } from 'sonner'
-import { useEffect } from 'react'
-import { usePathname } from 'next/navigation'
-import { Analytics } from '@vercel/analytics/react'
-import { QueryClientProvider } from '@tanstack/react-query'
-
-import { queryClient } from '@/services/react-query/query-client'
 import { Header } from '@/sections/header/header'
 import { Footer } from '@/sections/footer/footer'
-import { ThemeProvider } from '@/providers'
-
+import { ClientProviders, ScrollToTop, ThemeProvider } from '@/components/providers'
 import './globals.css'
 
-export default function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
-  const pathname = usePathname()
-
-  useEffect(() => {
-    window.scrollTo(0, 0)
-  }, [pathname])
-
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
-      <head>
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin="" />
-        <link href="https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&display=swap" rel="stylesheet" />
-      </head>
       <body className="antialiased dark:bg-[#000]">
-        <QueryClientProvider client={queryClient}>
-          <Toaster position="top-right" />
+        <ClientProviders>
           <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
+            <ScrollToTop />
             <Header />
             {children}
             <Footer />
           </ThemeProvider>
-          <Analytics />
-        </QueryClientProvider>
+        </ClientProviders>
       </body>
     </html>
   )
