+++ components/animation/timeline.tsx
@@ -1,8 +1,7 @@
 'use client'
+
 import React, { useEffect, useRef, useState } from 'react'
 import { useScroll, useTransform, motion } from 'framer-motion'
-
-import { Logo } from '../logo/logo'
 
 interface TimelineEntry {
   title: string
@@ -32,22 +31,20 @@ export const Timeline = ({ data }: { data: TimelineEntry[] }) => {
   const opacityTransform = useTransform(scrollYProgress, [0, 0.1], [0, 1])
 
   return (
-    <div className="container w-full font-sans" ref={containerRef}>
-      <div className="pt-20">
-        <h2 className="mb-4 flex max-w-4xl items-center gap-2 text-lg font-medium text-black dark:text-white md:text-4xl">
-          Latest updates of - <Logo className="text-2xl" />
-        </h2>
+    <div className="container w-full" ref={containerRef}>
+      <div className="mt-[50px]">
+        <h2 className="mb-4 flex max-w-4xl items-center gap-2 text-lg font-medium text-black dark:text-white md:text-4xl">Latest updates of - PortfoliosIo</h2>
       </div>
       <div ref={ref} className="relative pb-20">
         {data.map((item, index) => (
           <div key={index} className="flex justify-start pt-10 md:gap-10 md:pt-40">
             <div className="sticky top-40 z-40 flex max-w-xs flex-col items-center self-start md:w-full md:flex-row lg:max-w-sm">
-              <div className="absolute left-3 flex h-10 w-10 items-center justify-center rounded-full bg-white dark:bg-black md:left-3">
+              <div className="absolute -left-[18px] flex h-10 w-10 items-center justify-center rounded-full bg-white dark:bg-black">
                 <div className="h-4 w-4 rounded-full border border-neutral-300 bg-neutral-200 p-2 dark:border-neutral-700 dark:bg-neutral-800" />
               </div>
               <h3 className="hidden text-xl font-bold text-neutral-500 dark:text-neutral-500 md:block md:pl-20 md:text-5xl">{item.title}</h3>
             </div>
-            <div className="relative w-full pl-20 pr-4 md:pl-4">
+            <div className="relative w-full pl-20 md:pl-4">
               <h3 className="mb-4 block text-left text-2xl font-bold text-neutral-500 dark:text-neutral-500 md:hidden">{item.title}</h3>
               {item.content}{' '}
             </div>
@@ -57,7 +54,7 @@ export const Timeline = ({ data }: { data: TimelineEntry[] }) => {
           style={{
             height: height + 'px',
           }}
-          className="absolute left-8 top-0 w-[2px] overflow-hidden bg-[linear-gradient(to_bottom,var(--tw-gradient-stops))] from-transparent from-[0%] via-neutral-200 to-transparent to-[99%] [mask-image:linear-gradient(to_bottom,transparent_0%,black_10%,black_90%,transparent_100%)] dark:via-neutral-700 md:left-8"
+          className="absolute top-0 w-[2px] overflow-hidden bg-[linear-gradient(to_bottom,var(--tw-gradient-stops))] from-transparent from-[0%] via-neutral-200 to-transparent to-[99%] [mask-image:linear-gradient(to_bottom,transparent_0%,black_10%,black_90%,transparent_100%)] dark:via-neutral-700 md:left-0"
         >
           <motion.div
             style={{
