+++ components/providers/index.tsx
@@ -1,0 +1,3 @@
+export { ThemeProvider } from './theme-provider'
+export { ClientProviders } from './client-providers'
+export { ScrollToTop } from './scroll-to-top'
