--- providers/index.tsx
@@ -1,1 +1,0 @@
-export { ThemeProvider } from './theme-provider'
